Separating ToggleButton files.
