+++ android-design-system/src/main/java/com/adammcneilly/pocketleague/android/design/components/togglebutton/ToggleButtonOption.kt
@@ -1,0 +1,10 @@
+package com.adammcneilly.pocketleague.android.design.components.togglebutton
+
+/**
+ * Defines any option that can appear inside a [ToggleButtonRow].
+ *
+ * @param[text] The string that should be displayed to the user inside this button.
+ */
+data class ToggleButtonOption(
+    val text: String,
+)
