+++ android-design-system/src/main/java/com/adammcneilly/pocketleague/android/design/components/togglebutton/ToggleButtonRow.kt
@@ -1,65 +1,22 @@
-package com.adammcneilly.pocketleague.android.design.components
+package com.adammcneilly.pocketleague.android.design.components.togglebutton
 
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.text.intl.Locale
-import androidx.compose.ui.text.toUpperCase
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.zIndex
 import com.adammcneilly.pocketleague.android.design.theme.PocketLeagueTheme
 
 /**
- * Defines any option that can appear inside a [ToggleButtonRow].
- *
- * @param[text] The string that should be displayed to the user inside this button.
+ * TBD
  */
-data class ToggleButtonOption(
-    val text: String,
-)
-
-@Composable
-fun ToggleButton(
-    shape: Shape,
-    onClick: () -> Unit,
-    text: String,
-    selected: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    val colorsToUse = if (selected) {
-        ButtonDefaults.outlinedButtonColors(
-            containerColor = MaterialTheme.colorScheme.primary,
-            contentColor = MaterialTheme.colorScheme.onPrimary,
-        )
-    } else {
-        ButtonDefaults.outlinedButtonColors(
-            containerColor = MaterialTheme.colorScheme.background,
-            contentColor = MaterialTheme.colorScheme.onBackground,
-        )
-    }
-
-    OutlinedButton(
-        onClick = onClick,
-        shape = shape,
-        colors = colorsToUse,
-        modifier = modifier,
-    ) {
-        Text(
-            text = text.toUpperCase(Locale.current),
-        )
-    }
-}
-
 @Composable
 fun ToggleButtonRow(
     options: List<ToggleButtonOption>,
@@ -99,7 +56,31 @@ fun ToggleButtonRow(
                 onClick = {},
                 text = option.text,
                 selected = (option == selectedOption),
-                modifier = Modifier
+                modifier = when (index) {
+                    0 -> {
+                        if (selectedOption == option) {
+                            Modifier
+                                .offset(0.dp, 0.dp)
+                                .zIndex(1f)
+                        } else {
+                            Modifier
+                                .offset(0.dp, 0.dp)
+                                .zIndex(0f)
+                        }
+                    }
+                    else -> {
+                        val offset = -1 * index
+                        if (selectedOption == option) {
+                            Modifier
+                                .offset(offset.dp, 0.dp)
+                                .zIndex(1f)
+                        } else {
+                            Modifier
+                                .offset(offset.dp, 0.dp)
+                                .zIndex(0f)
+                        }
+                    }
+                }
                     .weight(1F),
             )
         }
@@ -120,13 +101,12 @@ private fun ToggleButtonRowPreview() {
     val past = ToggleButtonOption("Past")
 
     PocketLeagueTheme {
-        Surface(
-            modifier = Modifier
-                .padding(16.dp),
-        ) {
+        Surface {
             ToggleButtonRow(
                 options = listOf(upcoming, past),
                 selectedOption = upcoming,
+                modifier = Modifier
+                    .padding(16.dp),
             )
         }
     }
